QSet: add an rvalue overload of unite()
The old QSet::unite() overload cannot really be optimized, because the
RHS set was passed by cref, so cannot be modified, even if it were
already detached.

If we could, we could insert-or-overwrite from the LHS into the RHS
and then swap the two, if the RHS was larger than the LHS. This was
the original idea of 92acc94fa98d19929d28feb5d543acf8c50c2290, which,
however, ignored the implicit detach of the larger set if it was
shared.

So add a unite(QSet&&) overload that, when called, does have the
ability to change the RHS, incl. using it as the base for the
result. We just need a way to insert_or_assign into a QSet, which is
only possible using private API, because QHash never overwrites a
once-inserted key, and that's exactly what we need to perform this
operation while staying consistent with the QSet insertion behavior.

[ChangeLog][QtCore][QSet] Added unite(QSet&&) overload of unite(const
QSet&). Ditto for the `+` and `|` operators.

Task-number: QTBUG-132500
Change-Id: I31592d7bec3b0702e746b69e4f024708ad6a3cc4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>