Verify window's position is correct to fix flakiness when clicking
tst_QWindow::testInputEvents() is flaky where the simulated click is
sometimes not received by the window. This is caused by an incorrect
framePosition, where the position of the window is taken instead.

Remove the frame of the window and verify that the window's
position has been updated by using the ensurePositionTopLeft()
helper function.

Include QtTest/private/qtesthelpers_p.h to use the helper function.

Fix flaky tst_QWindow::testInputEvents on Opensuse

Change-Id: Ie65a70e7a478ff2db60a11f6ef6419dac3489ee6
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>