QSqlDatabase: fix historical mistake of defaultConnection being non-const
This was likely never intended, and was a mistake in the API design,
missing the second const after the pointer. It should have been:
    static const char *const defaultConnection;

This commit fixes that mistake by replacing the variable with a
constexpr one that can't be modified.

We can actually do better by moving the name to the header, so it will
get emitted in each library, application, or plugin instead of trying to
load from QtSql.

[ChangeLog][Potentially Source-Incompatible Changes] The static
QSqlDatabase::connectionName member variable has changed from a
non-const pointer to a constexpr array. Modifying the pointer was most
likely a mistake in code and would produce a misbehaving application.

Change-Id: Icf2eddc865c2d6dde0c7fffda06e55e215e5f26d
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>