QPicture: (almost) work around overflow in bounding box calculations
QPicture keeps a bounding box of its drawing, and updates it as it
processes paint commands, extending it if necessary. The problem is that
these calculations are prone to overflows:

* drawing (and thus the bounding box) is floating point based, but
  QPicture only stores a QRect. If the drawing's BB overflows the
  integer range, QPicture overflows.
* Since the stored QRect is `int` based (signed), if the drawing spans
  "too far" (e.g.  from INT_MIN/2 to INT_MAX/2), it'll again overflow
  the range. This impacts serialization, as the protocol serializes the
  bounding box's topleft+size.

For the former, clamp the qreal-based into the integer range before
the conversion.

For the latter, as a dirty work around, when calculating a new bounding
box, directly set the bounding box corners instead of topleft+size. This
is NOT A FIX by any means. I'm trading one source of UB for another; the
only objective is hiding an assertion from a test.

Some design change is needed to properly fix this (keeping in mind the
stability of the serialized format).

Task-number: QTBUG-133293
Change-Id: Ia741e1e73bdf7eee21f66b4ea5c1fae3430b468d
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>