QFutureInterface: mark the unused setContinuation() overloads as removed
This is a cleanup after the previous patch that introduced the new
setContinuation() overloads. The old API cannot be used correctly with
the new approach that requires to always set a continuation type, so
let's mark it as removed.

Technically, this change is SiC. However, QFutureInterfaceBase was
never documented as a public class, so users are not supposed to derive
from it and should not use setContinuation() methods directly.

Task-number: QTBUG-130662
Change-Id: Idf3d9f5de00a7c4ac34ee3dde1d58902e3a27904
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>