QNumeric: make the overflowing helpers constexpr
The qAdd/Sub/MulOverflow functions are useful to have in constexpr code.

In fact, they all already have some generic and constexpr-friendly
implementations (but see below).

The only issue is that these functions form a rather convoluted overload
set, which dispatch to different implementations (builtin, intrinsics)
depending on the compiler's capabilities.

Instead, centralize the implementation of the functions, which allows us
to detect if they are being called from constant evaluation and fall
back to the generic code in that case.

* On GCC, Clang, ICC-as-Clang: call the compiler builtin; it's constexpr
  on all the supported versions.

  In fact, make the detection more accurate: the old code excluded all
  Clang versions (incl. ICC-as-Clang) on 32-bit from using the builtins,
  including for addition and subtraction. That seems unnecessary,
  as they work just fine.

  The problem was that, for 64-bit arguments, Clang would generate a
  call to __mulodi4, and that failed to link. We can just special-case
  that; moreover, Clang 14 fixed this issue [1].

  Therefore, detect properly the compiler, incl. the Clang version.

* Otherwise, if we're in constant evaluation, call the generic
  implementation. This poses a problem: on platforms without native
  128-bit support, we are lacking an implementation of qMulOverflow
  when multiplying 64 bit quantities.

  In general, we did not guarantee its availability, but this regresses
  MSVC, which doesn't support __int128 but instead relied entirely
  on its intrinsics to perform 128-bit multiplications.

  To cope with this, add an implementation of a wide multiplication.

* Otherwise, use intrinsics (which are not constexpr).

* Otherwise, use the generic code.

There's a hair in this soup: INTEGRITY seems to only have non-constexpr
builtin (intrinsics), and it's also unable to detect if we're in
constant evaluation. This means that, during constant evaluation, we're
going to hit an error (call to the non-constexpr intrinsic).

Getting rid of the intrinsic and forcing the usage of the generic
version is tempting, but for now, these functions will simply not be
constexpr there.

A few drive-by fixes:

* The aforementioned detection of Clang, so it uses the intrinsics;
* Use Q_CC_GHS instead of Q_OS_INTEGRITY
* A couple of codepaths were checking if we are in two's complement,
  and doing different things otherwise. There is no need for the
  check, we never supported anything but two's complement.

[1] https://godbolt.org/z/v4z6W7qbd

[ChangeLog][QtCore][QtNumeric] The q*Overflow family of functions
are now callable from constant expressions on certain compilers.

Done-with: Thiago
Change-Id: I9fa2ab50a02b31358a916656e5b3329d77344efe
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>