tst_QShortcut: Test that modifier only shortcuts are not triggered
The QShortcutMap triggers shortcuts on key press. It's okay with normal
shortcuts but modifier only shortcuts require extra care. Depending on
the context, they can be triggered either on key press or key release.
For example, for push to talk, they should be triggered on key press,
but if a modifier only shortcut is assigned to a dashboard or something,
then it should be triggered on key release so the dashboard is not
accidentally opened when pressing another shortcut that starts with the
same modifier. The QShortcutMap currently doesn't provide support for
modifier only shortcuts.

The proposed new test case verifies that a modifier only shortcut will
not be accidentally triggered if there is a normal shortcut with the
same modifier keys.

Task-number: QTBUG-132435
Change-Id: I612d0239b29f8c1730016d10257def039b5e6cf1
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>