qobject_cast: optimize when the target is marked final
If it is, then we can perform a direct meta object pointer comparison.
Comparing meta objects is, after all, what QMetaObject::inherits() and
QMetaObject::cast()) are doing:

    const QMetaObject *m = this;
    do {
        if (metaObject == m)
            return true;
    } while ((m = m->d.superdata));

But if we know the class is final, we know an object can either be
exactly of the type we want or not be so at all.

Change-Id: If6333547d215c1d84d1cfffdfd82c04e75251608
Reviewed-by: Marc Mutz <marc.mutz@qt.io>