Long live QCheckedInt
This is a private and minimal "safe int" API for internal uses.
It has configurable policies for

* how to carry out a given arithmetic operation (check for overflow,
  just do the native language operation, ...);
* how to report an error in case the operation fails (Q_ASSERT, ignore
  it, ...).

By default, operations are implemented via the q*Overflow family of
functions, and Q_ASSERT is used to report errors. This means an
arithmetic error will be fatal in debug mode, and implement wraparound
in release mode. However, UB will be prevented; the compiler will
generate the same assembly, but it won't be allowed to exploit overflows
as an UB source (that's because q*Overflow has well-defined behavior).
In C++26 jargon, the program will have EB at this point.

The class is also supposed to be binary compatibile with the type it
wraps, so that it can be used to replace an `int` data member, without
breaking ABI. (Of course the mangling will be different, but this is an
implementation detail and it's not supposed to be exposed.)

It's woefully incomplete (operator% is missing, for instance); I just
needed a MVP to support the upcoming changes to the geomery classes.

This is meant to stay completely private, users that need a real
implementation should use Howard Hinnart's bbi.

Change-Id: I69ea20e872e395a505c5748f24105469dfb60d84
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>