iOS: Trigger expose on screen brightness change for windows that opt into EDR
When the screen brightness of the display changes, the active EDR headroom of
the display might also be updated.

We don't have a specific signal for these kinds of changes on a screen
or window level, so for now we do the next best thing, which is to trigger
an expose even for any relevant windows that might need to re-render to
reflect the updated EDR headroom.

Change-Id: I72ed3b8335ad763f66043daf6fcbac616425845b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>