macOS: Trigger expose on screen param change for windows that opt into EDR
When the active EDR headroom of a display is updated, due to its brightness
setting changing e.g., the system sends us a screen parameter change.

We don't have a specific signal for these kinds of changes on a screen
or window level, so for now we do the next best thing, which is to trigger
an expose even for any relevant windows that might need to re-render to
reflect the updated EDR headroom.

Change-Id: I1edd8ff6b9714b558a9ae92d45aaee56a838af7c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>