Geometry classes: honor the documentation when it talks about rounding
The floating-point based geometry classes promise rounding in their
documentation, in at least two places.

1) when fp math is applied to an integer-based class, the result is
   the same integer class. For instance, QPoint(1, 2) * 3.14 yields
   a QPoint (!). The documentation says that the result is rounded to
   the closest integer.

2) when a fp-based class is converted to the corresponding integer-based
   one (e.g. QRectF::toRect()). These functions have the same notice.

Insofar qRound has been used in the implementation of these functions.
There's a problem: qRound does not actually honor the "round to closest
integer" contract. Specifically, if the input floating point is out of
range for the destination type, then the behavior is undefined
(cf. a47a5932a64a8867077cab7c8efa57587b637481).

Since we have reports of cases where out-of-range values are indeed
passed through these functions, I'm extremely wary of changing the
behavior to be UB. Instead, I'm making these functions actually do what
the documentation says they do.

To this end, refactor the various qRound overloads and split the
rounding part from the casting to the destination type. The rounding is
kept as private API, using the existing arch-specific implementations.
The casting is generic and it's in the public qRound overloads.

Then, add a private qSaturateRound which re-uses the casting, clamps to
range ("saturating" -- which is still casting to the closest integer
value), and casts to the final type (only int, at the moment). This new
API is then used in the geometry classes when they round.

Fixes: QTBUG-133261
Change-Id: Iba0d9baa17c136a976a945742c117e4cfac520b3
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>